requirements.txt
python-dotenv==0.14.0
psycopg2-binary==2.8.5
psycopg2==2.8.5
Flask>=1.1.1
flask_cors>=3.0.8
